+++ src/os2datascanner/projects/report/reportapp/admin.py
@@ -4,7 +4,6 @@ from django.contrib.auth.admin import UserAdmin
 from django.contrib.auth.models import User
 
 from .models.roles.remediator import Remediator
-from .models.roles.dpo import DataProtectionOfficer
 from .models.roles.defaultrole import DefaultRole
 from .models.documentreport import DocumentReport
 
@@ -80,11 +79,6 @@ class DefaultRoleAdmin(admin.ModelAdmin):
 
 @admin.register(Remediator)
 class RemediatorAdmin(admin.ModelAdmin):
-    list_display = ('user', )
-
-
-@admin.register(DataProtectionOfficer)
-class DataProtectionOfficerAdmin(admin.ModelAdmin):
     list_display = ('user', )