+++ src/os2datascanner/projects/report/reportapp/management/commands/send_notifications.py
@@ -30,7 +30,7 @@ from os2datascanner.utils.system_utilities import time_now
 
 from ...models.roles.remediator import Remediator
 from ...views import statistics_views
-from ...utils import user_is_superadmin, user_is
+from ...utils import user_is
 from ...models.documentreport import DocumentReport
 from ...models.roles.defaultrole import DefaultRole
 from ....organizations.models.account import Account
@@ -235,7 +235,7 @@ class Command(BaseCommand):
 
         # If the user falls under either "superadmin" or remediator,
         # let the user know how many of these results stem from that.
-        if user_is_superadmin(user):
+        if user.is_superuser:
             superadmin_bound_results = context["superadmin_bound_results"] = user_results.filter(
                 only_notify_superadmin=True).count()
             total_result_count += superadmin_bound_results