+++ src/os2datascanner/projects/report/reportapp/views/support_views.py
@@ -1,7 +1,8 @@
 from django.views.generic import TemplateView
 
 from ..utils import convert_context_to_email_body
-from ..models.roles.dpo import DataProtectionOfficer
+from ...organizations.models.position import Position
+from os2datascanner.core_organizational_structure.models.position import Role
 
 
 class SupportButtonView(TemplateView):
@@ -9,6 +10,10 @@ class SupportButtonView(TemplateView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        context["dpo_contacts"] = DataProtectionOfficer.objects.filter(contact_person=True)
+        context["dpo_contacts"] = Position.objects.filter(
+            role=Role.DPO,
+            account__contact_person=True,
+            account__organization=self.request.user.account.organization
+        )
         context["email_body"] = convert_context_to_email_body(self.request)
         return context